Does now paint the "events" to a .png file inside ./res. And organized the project by moving classes/code
